requirements.txt
python-telegram-bot==13.7
python-dotenv==1.0.1
pytube==15.0.0
moviepy==1.0.3
pyTelegramBotAPI==4.15.4